Rollback the deploy instead of continuing when rsync fails
When rsync fails, I'm getting a deploy that looks like it went fine
except for an inconspicuous message saying that rsync failed and a
production deployment that's still on the old code version.  Sometimes I
don't notice right away that the new code is not actually live, and this
is a big problem.

I'd like to see a clear error message when rsync fails.

Raise an exception of rsync fails to any of the servers receiving a code
update.