Add the `build_script` concept from the `copy` strategy.
The basic idea is that there are some things we would prefer to do on
the local cache before `rsync`-ing it over to all the target machines.
In my particular case, this is a `git secret reveal`; doing it on the
local cache means I don't have to provision a private key on every
production web server. I can imagine there are other things that would
be useful, as well.